cleanup

+++ src/gamestate.h
@@ -282,16 +282,20 @@ bool g_add_default_action(gamestate* const g, entityid id, int action);
 bool g_set_default_action(gamestate* const g, entityid id, int action);
 entity_action_t g_get_default_action(const gamestate* const g, entityid id);
 
-entityid* g_sort_inventory(gamestate* const g, entityid* inventory, int inv_count, inventory_sort sort_type);
+//entityid* g_sort_inventory(gamestate* const g, entityid* inventory, int inv_count, inventory_sort sort_type);
+entityid* g_sort_inventory(gamestate* const g, entityid* inventory, size_t inv_count, inventory_sort sort_type);
+//entityid* g_sort_inventory(gamestate* const g, entityid* inventory, int inv_count, inventory_sort sort_type);
 bool g_has_inventory(const gamestate* const g, entityid id);
 bool g_add_inventory(gamestate* const g, entityid id);
 bool g_add_to_inventory(gamestate* const g, entityid id, entityid itemid);
 bool g_remove_from_inventory(gamestate* const g, entityid id, entityid itemid);
-entityid* g_get_inventory(const gamestate* const g, entityid id, int* count);
+//entityid* g_get_inventory(const gamestate* const g, entityid id, int* count);
+entityid* g_get_inventory(const gamestate* const g, entityid id, size_t* count);
 size_t g_get_inventory_count(const gamestate* const g, entityid id);
 bool g_has_item_in_inventory(const gamestate* const g, entityid id, entityid itemid);
 //bool g_update_inventory(gamestate* const g, entityid id, entityid* new_inventory);
-bool g_update_inventory(gamestate* const g, entityid id, entityid* new_inventory, int new_inventory_count);
+//bool g_update_inventory(gamestate* const g, entityid id, entityid* new_inventory, int new_inventory_count);
+bool g_update_inventory(gamestate* const g, entityid id, entityid* new_inventory, size_t new_inventory_count);
 
 bool g_has_target(const gamestate* const g, entityid id);
 bool g_add_target(gamestate* const g, entityid id, vec3 target);
